% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEM.R
\name{B_Complete}
\alias{B_Complete}
\title{information matrix}
\usage{
B_Complete(b_exp, b_out, se_exp, se_out, theta, r, c, p, tau_0, tau_1)
}
\description{
calculate the informationmatrix based on the complete data log-likelihood for standard error
}
\keyword{internal}
