% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEM.R
\name{CEM}
\alias{CEM}
\title{CEM algorithm}
\usage{
CEM(
  b_exp,
  b_out,
  se_exp,
  se_out,
  initial_theta,
  initial_r,
  initial_c,
  initial_p,
  EM_start,
  EM_maxit,
  maxit,
  n
)
}
\description{
Run CEM algorithm with one starting value
}
\keyword{internal}
