% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEM.R
\name{EM}
\alias{EM}
\title{EM algorithm}
\usage{
EM(
  b_exp,
  b_out,
  se_exp,
  se_out,
  initial_theta,
  initial_r,
  initial_c,
  initial_p,
  maxit,
  n
)
}
\description{
For starting the CEM algorithm
}
\keyword{internal}
