% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FG_T2D}
\alias{FG_T2D}
\title{Effect of Fast Glucose (FG) on Type 2 Diabetes (T2D)}
\format{
A \code{data.frame} with 17 rows and 10 variables.
}
\source{
\url{https://www.nature.com/articles/s41588-020-0631-4}
}
\usage{
FG_T2D
}
\description{
Effect of Fast Glucose (FG) on Type 2 Diabetes (T2D)
}
\keyword{datasets}
