% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixIE_MA.R
\name{mixIE_multiple_start}
\alias{mixIE_multiple_start}
\title{mixIE_multiple_start}
\usage{
mixIE_multiple_start(
  b_exp,
  b_out,
  se_exp,
  se_out,
  n,
  initial_theta_n = 50,
  initial_r = 0,
  initial_c = 1,
  initial_p = 0.2,
  flip = 1,
  EM_start = T,
  EM_maxit = 2,
  maxit = 200,
  ivw = T,
  egger = T,
  lb.theta = NULL,
  ub.theta = NULL
)
}
\arguments{
\item{b_exp}{A vector of SNP effects on the exposure variable, usually obtained from a GWAS.}

\item{b_out}{A vector of SNP effects on the outcome variable, usually obtained from a GWAS.}

\item{se_exp}{A vector of standard errors of \code{b_exp}.}

\item{se_out}{A vector of standard errors of \code{b_out}.}

\item{n}{Sample size of either one of the GWAS dataset.}

\item{initial_theta_n}{Number of different thetas generated, default is 50.}

\item{initial_r}{Initial value of r, the averaged pleiotropic effect, default is 0.}

\item{initial_c}{Initial value of c, the overdispersion parameter for invalid IVs, default is 1.}

\item{initial_p}{Initial value of the proportion of invalid IVs, default is 0.2.}

\item{flip}{Whether to reorient the SNPs like Egger regression?}

\item{EM_start}{Whether to use EM algorithm to start the CEM? Default is TRUE.}

\item{EM_maxit}{Number of iterations used in EM algorithm to start the CEM, default is 2.}

\item{maxit}{Maximum number of iterations for each optimization, default is 200.}

\item{ivw}{Whether to add the fixed effect IVW in the model candidates explictly? Default is TRUE.}

\item{egger}{Whether to add the Egger regression in the model candidates explictly? Default is TRUE.}

\item{lb.theta}{Lower bound of theta from which starting values are generated.}

\item{ub.theta}{Upper bound of theta from which starting values are generated.}
}
\value{
A list
\describe{
\item{theta}{A vector of estimated causal effect for different starting values ordered by BIC}
\item{se}{A vector of corresponding standard error of \code{theta}}
\item{pval}{A vector of corresponding two-sided p-value of \code{theta}}
\item{c}{A vector of estimated overdispersion parameter for different starting values ordered by BIC}
\item{r}{A vector of estimated pleiotropic effect for different starting values ordered by BIC}
\item{ser}{A vector of corresponding standard error of \code{r}}
\item{p}{A vector of estimated proportion of invalid IVs for different starting values ordered by BIC}
\item{BIC}{A vector of BIC for different starting values sorting increasingly}
\item{niter}{A vector of number of iterations for different starting values ordered by BIC}
\item{tau_1_mat}{A matrix of posterior probabilities of each IVs being invalid for different starting values ordered by BIC}
}
}
\description{
Perform CEM algorithm with multiple starting values
}
