% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEM.R
\name{S_Complete}
\alias{S_Complete}
\title{gradient}
\usage{
S_Complete(b_exp, b_out, se_exp, se_out, theta, r, c, p, tau_0, tau_1)
}
\description{
calculate the gradient based on the complete data log-likelihood for standard error
}
\keyword{internal}
